backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
alembic==1.14.0
yfinance==0.2.51
requests==2.32.3
pydantic==2.10.3
python-dotenv==1.0.1
apscheduler==3.10.4
